requirements.txt
numpy==1.23.1
pandas==1.5.0
scikit-learn==1.1.2
tensorflow==2.10.0
pytorch==1.12.1
transformers==4.21.2
fastapi==0.85.0
uvicorn==0.20.0
pydantic==1.10.2
opencv-python-headless==4.6.0.66
iot-hub-device-client==2.0.0
